Add files via upload
Newer iOS updates have a different version of Safari, so I had to add instructions for locating the share button on the newer version. I can't seem to find a good solution
to not creating a new line when using <strong> in the step 2 paragraph. Makes the instructions a little more confusing than it should be, but for now it seems to work fine.