Add files via upload
This time I'm trying to hold the user's hand in the guide since after a couple test runs the steps weren't clear enough which was understandable since one of the steps was to click the blue button, but there was no blue button until you finished the guide which wasn't stated anywhere. I'm hoping this version is completely dummy proof, but my Dad may give me a run for my money.