Add files via upload
Well, my Dad has officially given me a run for my money and now I have to ensure that it is as easy as humanly possible to ensure that the instructions are clear. This version 
holds the user's hand and does a decent job at ensuring that the instructions are read from top to bottom.